+++ src/components/form/fields/dateField.jsx
@@ -22,7 +22,7 @@ const InputCol = styled.div`
     flex-basis: ${props => props.width}%;
 `;
 
-const DateField = ({ i18n, inputId, label, locale, onChange, path, required, showLabel, validation, value }) => {
+const DateField = ({ i18n, inputId, label, locale, onChange, path, required, showLabel, validation={}, value }) => {
     const [day, setDay] = useState(isValued(value) ? value.raw.day : undefined)
     const [month, setMonth] = useState(isValued(value) ? value.raw.month : undefined)
     const [year, setYear] = useState(isValued(value) ? value.raw.year : undefined)
@@ -159,7 +159,7 @@ export default function dateField(props, config) {
                 const dt = value ? DateTime.fromISO(value) : DateTime.invalid('empty value')
                 return dt.isValid ? { raw: dt } : undefined
             },
-            unbind: (value) => value.raw.toISODate()
+            unbind: (value) => value && value.raw.toISODate()
         },
         validator: props.validator ? datetimeValidator(config.language).and(props.validator) : datetimeValidator(config.language),
         component: DateField,