+++ src/components/widget/widget.tsx
@@ -7,7 +7,7 @@ import type { I18nMessages } from '../../core/i18n';
 
 import { ConfigProvider } from '../../contexts/config';
 import { I18nProvider } from '../../contexts/i18n'
-import { ReacfiveProvider } from '../../contexts/reachfive'
+import { ReachfiveProvider } from '../../contexts/reachfive'
 import { RoutingProvider } from '../../contexts/routing'
 import { SessionProvider } from '../../contexts/session'
 
@@ -32,7 +32,7 @@ export type Context = {
 type PrepareFn<P, U> = (options: PropsWithI18n<PropsWithTheme<P>>, context: Context) => PropsWithI18n<PropsWithTheme<U>> | PromiseLike<PropsWithI18n<PropsWithTheme<U>>>
 
 type CreateWidget<P, U> = {
-    component: ComponentType<U>
+    component: ComponentType<Omit<U, 'theme'>>
     prepare?: PrepareFn<P, U>
 } & WidgetContainerProps
 
@@ -42,24 +42,24 @@ export function createWidget<P, U = P>({
     ...widgetAttrs
 }: CreateWidget<P, U>) {
     return (options: PropsWithTheme<PropsWithI18n<P>>, context: Context) => {
-        return Promise.resolve(prepare(options, context)).then(preparedOptions => {
+        return Promise.resolve(prepare(options, context)).then(({ theme: customTheme, ...preparedOptions }) => {
             const Component = component
 
-            const theme: Theme = buildTheme(options.theme)
+            const theme: Theme = buildTheme(customTheme)
 
             return (
                 <ConfigProvider config={context.config}>
-                    <ReacfiveProvider client={context.apiClient}>
+                    <ReachfiveProvider client={context.apiClient}>
                         <SessionProvider session={context.session}>
                             <ThemeProvider theme={theme}>
-                                <I18nProvider defaultMessages={context.defaultI18n} messages={options.i18n}>
+                                <I18nProvider defaultMessages={context.defaultI18n} messages={preparedOptions.i18n}>
                                     <WidgetContainer {...widgetAttrs}>
                                         <Component {...preparedOptions} />
                                     </WidgetContainer>
                                 </I18nProvider>
                             </ThemeProvider>
                         </SessionProvider>
-                    </ReacfiveProvider>
+                    </ReachfiveProvider>
                 </ConfigProvider>
             );
         });
@@ -77,8 +77,8 @@ export function createMultiViewWidget<P, U = P>({ prepare, ...params }: MultiVie
 }
 
 export interface MultiViewWidgetProps<P, U> {
-    initialView: ((props: U) => string) | string
-    views: Record<string, ComponentType<U>>
+    initialView: ((props: Omit<U, 'theme'>) => string) | string
+    views: Record<string, ComponentType<Omit<U, 'theme'>>>
     initialState?: MultiWidgetState
     prepare?: PrepareFn<P, U>
 }
@@ -88,13 +88,13 @@ export type MultiWidgetState = Record<string, unknown> & {
 }
 
 function multiViewWidget<P, U>({ initialView, views, initialState = {} as MultiWidgetState }: MultiViewWidgetProps<P, U>) {
-    return class MultiViewWidget extends React.Component<U, MultiWidgetState> {
+    return class MultiViewWidget extends React.Component<Omit<U, 'theme'>, MultiWidgetState> {
 
         state = {
             ...initialState,
             activeView: typeof initialView === 'function' ? initialView(this.props) : initialView
         };
-        
+
         // _goTo = <View extends keyof typeof views, S extends ComponentProps<(typeof views)[View]>>(view: View, props?: S) => this.setState({
         _goTo = <S extends Record<string, unknown>>(view: string, params?: S) => this.setState({
             activeView: view as MultiWidgetState['activeView'],