+++ tests/widgets/socialAccounts/socialAccountsWidget.test.js
@@ -3,7 +3,7 @@
  */
 
 import { describe, expect, jest, test } from '@jest/globals';
-import { render, screen, waitFor } from '@testing-library/react'
+import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import '@testing-library/jest-dom'
 import 'jest-styled-components';
 
@@ -28,13 +28,15 @@ describe('Snapshot', () => {
             { config, apiClient }
         )
 
-        const { container, rerender } = await render(widget);
+        await waitFor(async () => {
+            const { container, rerender } = await render(widget);
 
-        await waitFor(() => expect(apiClient.getUser).toHaveBeenCalled())
-
-        await rerender(widget)
-        
-        expect(container).toMatchSnapshot();
+            await waitFor(() => expect(apiClient.getUser).toHaveBeenCalled())
+    
+            await rerender(widget)
+            
+            expect(container).toMatchSnapshot();
+        })
     };
 
     test('basic', generateSnapshot({
@@ -45,11 +47,13 @@ describe('Snapshot', () => {
 })
 
 describe('DOM testing', () => {
+    const handleUnlink = jest.fn().mockReturnValueOnce(Promise.resolve())
+
     const generateComponent = async ({ options = {}, config = defaultConfig, socialIdentities }) => {
 
         const apiClient = {
             getUser: jest.fn().mockReturnValueOnce(Promise.resolve({ socialIdentities })),
-            unlink: jest.fn(),
+            unlink: handleUnlink,
             on: jest.fn(),
             off: jest.fn(),
         };
@@ -59,11 +63,13 @@ describe('DOM testing', () => {
             { config, apiClient }
         )
 
-        const { rerender } = await render(widget);
+        await waitFor(async () => {
+            const { rerender } = await render(widget);
 
-        await waitFor(() => expect(apiClient.getUser).toHaveBeenCalled())
+            await waitFor(() => expect(apiClient.getUser).toHaveBeenCalled())
 
-        return await rerender(widget)
+            await rerender(widget)
+        })
     };
 
     describe('with default config', () => {
@@ -106,6 +112,30 @@ describe('DOM testing', () => {
             expect(screen.queryByTestId('identity-google')).toBeInTheDocument();
             expect(screen.queryByTestId('identity-line')).toBeInTheDocument();
             expect(screen.queryByText('socialAccounts.linkNewAccount')).not.toBeInTheDocument();
+        })
+
+        test('unlink identity', async () => {
+            await generateComponent({
+                socialIdentities: [
+                    { id: '123456789', provider: 'facebook', name: 'John Doe' },
+                    { id: '987654321', provider: 'google', name: 'John Doe' },
+                    { id: '000000000', provider: 'line', name: 'John Doe' },
+                ]
+            });
+
+            expect(screen.queryByTestId('identity-google')).toBeInTheDocument();
+            
+            const unlinkBtn = screen.queryByTestId('identity-google-unlink');
+            expect(unlinkBtn).toBeInTheDocument();
+            
+            fireEvent.click(unlinkBtn)
+
+            expect(handleUnlink).toHaveBeenCalledTimes(1)
+            expect(handleUnlink.mock.calls[handleUnlink.mock.calls.length - 1][0].identityId).toBe('987654321') // Google
+
+            await waitFor(() => {
+                expect(screen.queryByTestId('identity-google')).not.toBeInTheDocument();
+            })
         })
     })
 })