+++ CHANGELOG.md
@@ -7,7 +7,43 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.25.3] - 2023-03-08
+
+### Fixes
+
+- Make `@rollup/rollup-linux-x64-gnu` dependency optional
+- Fix React hook warning due to mismatch of React versions
+- Resolve not wrapped in act warning issue in tests
+- Fix custom theme option forwarding
+- Prevent phone-number field validation on empty value
+
+## [1.25.2] - 2023-02-21
+
+### Fixes
+
+- Fix types definitions bundling
+
+## [1.25.1] - 2023-02-19
+
+### Fixes
+
+- Fix bundle dependencies embeding
+
+## [1.25.0] - 2023-02-19
+
+### Changed
+
+- TypeScript migration
+
+### Fixes
+
+- Fix dateField validation when value is empty
+- Show API error when trying to unlink a social account with a non-fresh access_token
+
 ## [1.24.0] - 2023-01-05
+
+## Added
+
 - Add variable locale
 
 ## [1.23.1] - 2023-12-12
@@ -389,7 +425,15 @@ The eye icon is now correctly displayed in the Auth widget.
 
 First version of the SDK Web UI.
 
-[Unreleased]: https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.24.0...HEAD
+[Unreleased]: https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.25.3...HEAD
+
+[1.25.3]:  https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.25.2...v1.25.3
+
+[1.25.2]:  https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.25.1...v1.25.2
+
+[1.25.1]:  https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.25.0...v1.25.1
+
+[1.25.0]:  https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.24.0...v1.25.0
 
 [1.24.0]:  https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.23.1...v1.24.0